optimize RAM OOM issue during evaluation fix wrong dtype in loss

+++ model/softgroup.py
@@ -195,7 +195,8 @@ class SoftGroup(pl.LightningModule):
             labels = fg_instance_cls.new_full((fg_ious_on_cluster.size(0),), self.instance_classes)
             labels[pos_inds] = fg_instance_cls[pos_gt_inds]
             classification_criterion = ClassificationLoss()
-            classification_loss = classification_criterion(output_dict["cls_scores"], labels.long())
+            labels = labels.long()
+            classification_loss = classification_criterion(output_dict["cls_scores"], labels)
             losses["classification_loss"] = classification_loss
 
             """mask scoring loss"""
